Revert "Support formatting raw value on commit"
It was a very leaky abstraction: whenever state updated without being driven by
a `commit()` from the input view, it was difficult to make value formatting
happen.

Also, it's a bad assumption that every kind of "formatting" means a modification
of the raw value. Formatting is a pretty arbitrary but view-specific concern,
whose only state tie-in is that it should only happen at times determined by the
state.

(bonus points for async validation being a pain in the ass forcing better
design: it's tricky for input components to determine this state for themselves,
because simply formatting at commit time gets overwritten when validation
settles.)

Also hooking callbacks through the whole state system gets a little annoyingly
complex.

This reverts commit a7a6cca0752ba59d04d6fb6d6f0c89ad65d96dde.